Event dispatching works properly.
Removed redundant behavior of the window having an event callback that calls tc_app_on_event.
Events are now created in the sokol event callback and passed directly to tc_app_on_event.
Sokol logs are now passed to a callback to be handled by the engines own logging behaviour.